Rename calc.py to calculator.py
